move interceptors to robot client
